+++ node/npm/glide/index.ts
@@ -74,8 +74,8 @@ function loadNativeBinding() {
 function initialize() {
     const nativeBinding = loadNativeBinding();
     const {
-        BitOffsetOptions,
         BitmapIndexType,
+        BitOffsetOptions,
         BitwiseOperation,
         ConditionalChange,
         GeoAddOptions,
@@ -92,6 +92,7 @@ function initialize() {
         PeriodicChecks,
         Logger,
         LPosOptions,
+        ListDirection,
         ExpireOptions,
         FlushMode,
         GeoUnit,
@@ -100,6 +101,7 @@ function initialize() {
         SetOptions,
         ZaddOptions,
         ScoreBoundry,
+        UpdateOptions,
         RangeByIndex,
         RangeByScore,
         RangeByLex,
@@ -129,8 +131,8 @@ function initialize() {
     } = nativeBinding;
 
     module.exports = {
-        BitOffsetOptions,
         BitmapIndexType,
+        BitOffsetOptions,
         BitwiseOperation,
         ConditionalChange,
         GeoAddOptions,
@@ -147,6 +149,7 @@ function initialize() {
         PeriodicChecks,
         Logger,
         LPosOptions,
+        ListDirection,
         ExpireOptions,
         FlushMode,
         GeoUnit,
@@ -155,6 +158,7 @@ function initialize() {
         SetOptions,
         ZaddOptions,
         ScoreBoundry,
+        UpdateOptions,
         RangeByIndex,
         RangeByScore,
         RangeByLex,