+++ node/tests/SharedTests.ts
@@ -10,31 +10,28 @@
 import { expect, it } from "@jest/globals";
 import { v4 as uuidv4 } from "uuid";
 import {
+    BitmapIndexType,
     BitwiseOperation,
     ClosingError,
+    ConditionalChange,
     ExpireOptions,
+    FlushMode,
     GeoUnit,
+    GeospatialData,
     GlideClient,
     GlideClusterClient,
     InfoOptions,
     InsertPosition,
+    ListDirection,
     ProtocolVersion,
     RequestError,
     ScoreFilter,
     Script,
+    UpdateByScore,
     parseInfoResponse,
 } from "../";
 import { RedisCluster } from "../../utils/TestUtils";
 import { SingleNodeRoute } from "../build-ts/src/GlideClusterClient";
-import {
-    BitOffsetOptions,
-    BitmapIndexType,
-} from "../build-ts/src/commands/BitOffsetOptions";
-import { ConditionalChange } from "../build-ts/src/commands/ConditionalChange";
-import { FlushMode } from "../build-ts/src/commands/FlushMode";
-import { LPosOptions } from "../build-ts/src/commands/LPosOptions";
-import { GeoAddOptions } from "../build-ts/src/commands/geospatial/GeoAddOptions";
-import { GeospatialData } from "../build-ts/src/commands/geospatial/GeospatialData";
 import {
     Client,
     GetAndSetRandomValue,
@@ -284,8 +281,6 @@ export function runBaseTests<Context>(config: {
                 await client.set("foo", "bar");
                 const oldResult = await client.info([InfoOptions.Commandstats]);
                 const oldResultAsString = intoString(oldResult);
-                console.log(oldResult);
-                console.log(oldResultAsString);
                 expect(oldResultAsString).toContain("cmdstat_set");
                 checkSimple(await client.configResetStat()).toEqual("OK");
 
@@ -655,6 +650,131 @@ export function runBaseTests<Context>(config: {
                 await expect(client.setbit(setKey, 0, 0)).rejects.toThrow(
                     RequestError,
                 );
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `bitpos and bitposInterval test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient, cluster) => {
+                const key = `{key}-${uuidv4()}`;
+                const nonExistingKey = `{key}-${uuidv4()}`;
+                const setKey = `{key}-${uuidv4()}`;
+                const value = "?f0obar"; // 00111111 01100110 00110000 01101111 01100010 01100001 01110010
+
+                checkSimple(await client.set(key, value)).toEqual("OK");
+                expect(await client.bitpos(key, 0)).toEqual(0);
+                expect(await client.bitpos(key, 1)).toEqual(2);
+                expect(await client.bitpos(key, 1, 1)).toEqual(9);
+                expect(await client.bitposInterval(key, 0, 3, 5)).toEqual(24);
+
+                // -1 is returned if start > end
+                expect(await client.bitposInterval(key, 0, 1, 0)).toEqual(-1);
+
+                // `BITPOS` returns -1 for non-existing strings
+                expect(await client.bitpos(nonExistingKey, 1)).toEqual(-1);
+                expect(
+                    await client.bitposInterval(nonExistingKey, 1, 3, 5),
+                ).toEqual(-1);
+
+                // invalid argument - bit value must be 0 or 1
+                await expect(client.bitpos(key, 2)).rejects.toThrow(
+                    RequestError,
+                );
+                await expect(
+                    client.bitposInterval(key, 2, 3, 5),
+                ).rejects.toThrow(RequestError);
+
+                // key exists, but it is not a string
+                expect(await client.sadd(setKey, ["foo"])).toEqual(1);
+                await expect(client.bitpos(setKey, 1)).rejects.toThrow(
+                    RequestError,
+                );
+                await expect(
+                    client.bitposInterval(setKey, 1, 1, -1),
+                ).rejects.toThrow(RequestError);
+
+                if (cluster.checkIfServerVersionLessThan("7.0.0")) {
+                    await expect(
+                        client.bitposInterval(
+                            key,
+                            1,
+                            1,
+                            -1,
+                            BitmapIndexType.BYTE,
+                        ),
+                    ).rejects.toThrow(RequestError);
+                    await expect(
+                        client.bitposInterval(
+                            key,
+                            1,
+                            1,
+                            -1,
+                            BitmapIndexType.BIT,
+                        ),
+                    ).rejects.toThrow(RequestError);
+                } else {
+                    expect(
+                        await client.bitposInterval(
+                            key,
+                            0,
+                            3,
+                            5,
+                            BitmapIndexType.BYTE,
+                        ),
+                    ).toEqual(24);
+                    expect(
+                        await client.bitposInterval(
+                            key,
+                            1,
+                            43,
+                            -2,
+                            BitmapIndexType.BIT,
+                        ),
+                    ).toEqual(47);
+                    expect(
+                        await client.bitposInterval(
+                            nonExistingKey,
+                            1,
+                            3,
+                            5,
+                            BitmapIndexType.BYTE,
+                        ),
+                    ).toEqual(-1);
+                    expect(
+                        await client.bitposInterval(
+                            nonExistingKey,
+                            1,
+                            3,
+                            5,
+                            BitmapIndexType.BIT,
+                        ),
+                    ).toEqual(-1);
+
+                    // -1 is returned if the bit value wasn't found
+                    expect(
+                        await client.bitposInterval(
+                            key,
+                            1,
+                            -1,
+                            -1,
+                            BitmapIndexType.BIT,
+                        ),
+                    ).toEqual(-1);
+
+                    // key exists, but it is not a string
+                    await expect(
+                        client.bitposInterval(
+                            setKey,
+                            1,
+                            1,
+                            -1,
+                            BitmapIndexType.BIT,
+                        ),
+                    ).rejects.toThrow(RequestError);
+                }
             }, protocol);
         },
         config.timeout,
@@ -1099,6 +1219,116 @@ export function runBaseTests<Context>(config: {
                         "Operation against a key holding the wrong kind of value",
                     );
                 }
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `lmove list_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient, cluster) => {
+                if (cluster.checkIfServerVersionLessThan("6.2.0")) {
+                    return;
+                }
+
+                const key1 = "{key}-1" + uuidv4();
+                const key2 = "{key}-2" + uuidv4();
+                const lpushArgs1 = ["2", "1"];
+                const lpushArgs2 = ["4", "3"];
+
+                // Initialize the tests
+                expect(await client.lpush(key1, lpushArgs1)).toEqual(2);
+                expect(await client.lpush(key2, lpushArgs2)).toEqual(2);
+
+                // Move from LEFT to LEFT
+                checkSimple(
+                    await client.lmove(
+                        key1,
+                        key2,
+                        ListDirection.LEFT,
+                        ListDirection.LEFT,
+                    ),
+                ).toEqual("1");
+
+                // Move from LEFT to RIGHT
+                checkSimple(
+                    await client.lmove(
+                        key1,
+                        key2,
+                        ListDirection.LEFT,
+                        ListDirection.RIGHT,
+                    ),
+                ).toEqual("2");
+
+                checkSimple(await client.lrange(key2, 0, -1)).toEqual([
+                    "1",
+                    "3",
+                    "4",
+                    "2",
+                ]);
+                checkSimple(await client.lrange(key1, 0, -1)).toEqual([]);
+
+                // Move from RIGHT to LEFT - non-existing destination key
+                checkSimple(
+                    await client.lmove(
+                        key2,
+                        key1,
+                        ListDirection.RIGHT,
+                        ListDirection.LEFT,
+                    ),
+                ).toEqual("2");
+
+                // Move from RIGHT to RIGHT
+                checkSimple(
+                    await client.lmove(
+                        key2,
+                        key1,
+                        ListDirection.RIGHT,
+                        ListDirection.RIGHT,
+                    ),
+                ).toEqual("4");
+
+                checkSimple(await client.lrange(key2, 0, -1)).toEqual([
+                    "1",
+                    "3",
+                ]);
+                checkSimple(await client.lrange(key1, 0, -1)).toEqual([
+                    "2",
+                    "4",
+                ]);
+
+                // Non-existing source key
+                expect(
+                    await client.lmove(
+                        "{key}-non_existing_key" + uuidv4(),
+                        key1,
+                        ListDirection.LEFT,
+                        ListDirection.LEFT,
+                    ),
+                ).toEqual(null);
+
+                // Non-list source key
+                const key3 = "{key}-3" + uuidv4();
+                checkSimple(await client.set(key3, "value")).toEqual("OK");
+                await expect(
+                    client.lmove(
+                        key3,
+                        key1,
+                        ListDirection.LEFT,
+                        ListDirection.LEFT,
+                    ),
+                ).rejects.toThrow(RequestError);
+
+                // Non-list destination key
+                await expect(
+                    client.lmove(
+                        key1,
+                        key3,
+                        ListDirection.LEFT,
+                        ListDirection.LEFT,
+                    ),
+                ).rejects.toThrow(RequestError);
             }, protocol);
         },
         config.timeout,
@@ -2216,25 +2446,27 @@ export function runBaseTests<Context>(config: {
                 const membersScores = { one: 1, two: 2, three: 3 };
                 expect(
                     await client.zadd(key, membersScores, {
-                        conditionalChange: "onlyIfExists",
+                        conditionalChange: ConditionalChange.ONLY_IF_EXISTS,
                     }),
                 ).toEqual(0);
 
                 expect(
                     await client.zadd(key, membersScores, {
-                        conditionalChange: "onlyIfDoesNotExist",
+                        conditionalChange:
+                            ConditionalChange.ONLY_IF_DOES_NOT_EXIST,
                     }),
                 ).toEqual(3);
 
                 expect(
                     await client.zaddIncr(key, "one", 5.0, {
-                        conditionalChange: "onlyIfDoesNotExist",
+                        conditionalChange:
+                            ConditionalChange.ONLY_IF_DOES_NOT_EXIST,
                     }),
                 ).toEqual(null);
 
                 expect(
                     await client.zaddIncr(key, "one", 5.0, {
-                        conditionalChange: "onlyIfExists",
+                        conditionalChange: ConditionalChange.ONLY_IF_EXISTS,
                     }),
                 ).toEqual(6.0);
             }, protocol);
@@ -2254,27 +2486,27 @@ export function runBaseTests<Context>(config: {
 
                 expect(
                     await client.zadd(key, membersScores, {
-                        updateOptions: "scoreGreaterThanCurrent",
+                        updateOptions: UpdateByScore.GREATER_THAN,
                         changed: true,
                     }),
                 ).toEqual(1);
 
                 expect(
                     await client.zadd(key, membersScores, {
-                        updateOptions: "scoreLessThanCurrent",
+                        updateOptions: UpdateByScore.LESS_THAN,
                         changed: true,
                     }),
                 ).toEqual(0);
 
                 expect(
                     await client.zaddIncr(key, "one", -3.0, {
-                        updateOptions: "scoreLessThanCurrent",
+                        updateOptions: UpdateByScore.LESS_THAN,
                     }),
                 ).toEqual(7.0);
 
                 expect(
                     await client.zaddIncr(key, "one", -3.0, {
-                        updateOptions: "scoreGreaterThanCurrent",
+                        updateOptions: UpdateByScore.GREATER_THAN,
                     }),
                 ).toEqual(null);
             }, protocol);
@@ -4354,44 +4586,32 @@ export function runBaseTests<Context>(config: {
 
                 // simplest case
                 expect(await client.lpos(key, "a")).toEqual(0);
-                expect(
-                    await client.lpos(key, "b", new LPosOptions({ rank: 2 })),
-                ).toEqual(5);
+                expect(await client.lpos(key, "b", { rank: 2 })).toEqual(5);
 
                 // element doesn't exist
                 expect(await client.lpos(key, "e")).toBeNull();
 
                 // reverse traversal
-                expect(
-                    await client.lpos(key, "b", new LPosOptions({ rank: -2 })),
-                ).toEqual(2);
+                expect(await client.lpos(key, "b", { rank: -2 })).toEqual(2);
 
                 // unlimited comparisons
                 expect(
-                    await client.lpos(
-                        key,
-                        "a",
-                        new LPosOptions({ rank: 1, maxLength: 0 }),
-                    ),
+                    await client.lpos(key, "a", { rank: 1, maxLength: 0 }),
                 ).toEqual(0);
 
                 // limited comparisons
                 expect(
-                    await client.lpos(
-                        key,
-                        "c",
-                        new LPosOptions({ rank: 1, maxLength: 2 }),
-                    ),
+                    await client.lpos(key, "c", { rank: 1, maxLength: 2 }),
                 ).toBeNull();
 
                 // invalid rank value
                 await expect(
-                    client.lpos(key, "a", new LPosOptions({ rank: 0 })),
+                    client.lpos(key, "a", { rank: 0 }),
                 ).rejects.toThrow(RequestError);
 
                 // invalid maxlen value
                 await expect(
-                    client.lpos(key, "a", new LPosOptions({ maxLength: -1 })),
+                    client.lpos(key, "a", { maxLength: -1 }),
                 ).rejects.toThrow(RequestError);
 
                 // non-existent key
@@ -4407,45 +4627,29 @@ export function runBaseTests<Context>(config: {
 
                 // invalid count value
                 await expect(
-                    client.lpos(key, "a", new LPosOptions({ count: -1 })),
+                    client.lpos(key, "a", { count: -1 }),
                 ).rejects.toThrow(RequestError);
 
                 // with count
+                expect(await client.lpos(key, "a", { count: 2 })).toEqual([
+                    0, 1,
+                ]);
+                expect(await client.lpos(key, "a", { count: 0 })).toEqual([
+                    0, 1, 4,
+                ]);
                 expect(
-                    await client.lpos(key, "a", new LPosOptions({ count: 2 })),
-                ).toEqual([0, 1]);
-                expect(
-                    await client.lpos(key, "a", new LPosOptions({ count: 0 })),
+                    await client.lpos(key, "a", { rank: 1, count: 0 }),
                 ).toEqual([0, 1, 4]);
                 expect(
-                    await client.lpos(
-                        key,
-                        "a",
-                        new LPosOptions({ rank: 1, count: 0 }),
-                    ),
-                ).toEqual([0, 1, 4]);
-                expect(
-                    await client.lpos(
-                        key,
-                        "a",
-                        new LPosOptions({ rank: 2, count: 0 }),
-                    ),
+                    await client.lpos(key, "a", { rank: 2, count: 0 }),
                 ).toEqual([1, 4]);
                 expect(
-                    await client.lpos(
-                        key,
-                        "a",
-                        new LPosOptions({ rank: 3, count: 0 }),
-                    ),
+                    await client.lpos(key, "a", { rank: 3, count: 0 }),
                 ).toEqual([4]);
 
                 // reverse traversal
                 expect(
-                    await client.lpos(
-                        key,
-                        "a",
-                        new LPosOptions({ rank: -1, count: 0 }),
-                    ),
+                    await client.lpos(key, "a", { rank: -1, count: 0 }),
                 ).toEqual([4, 1, 0]);
             }, protocol);
         },
@@ -4511,18 +4715,15 @@ export function runBaseTests<Context>(config: {
                 checkSimple(await client.set(key1, value)).toEqual("OK");
                 expect(await client.bitcount(key1)).toEqual(26);
                 expect(
-                    await client.bitcount(key1, new BitOffsetOptions(1, 1)),
+                    await client.bitcount(key1, { start: 1, end: 1 }),
                 ).toEqual(6);
                 expect(
-                    await client.bitcount(key1, new BitOffsetOptions(0, -5)),
+                    await client.bitcount(key1, { start: 0, end: -5 }),
                 ).toEqual(10);
                 // non-existing key
                 expect(await client.bitcount(uuidv4())).toEqual(0);
                 expect(
-                    await client.bitcount(
-                        uuidv4(),
-                        new BitOffsetOptions(5, 30),
-                    ),
+                    await client.bitcount(uuidv4(), { start: 5, end: 30 }),
                 ).toEqual(0);
                 // key exists, but it is not a string
                 expect(await client.sadd(key2, [value])).toEqual(1);
@@ -4530,53 +4731,60 @@ export function runBaseTests<Context>(config: {
                     RequestError,
                 );
                 await expect(
-                    client.bitcount(key2, new BitOffsetOptions(1, 1)),
+                    client.bitcount(key2, { start: 1, end: 1 }),
                 ).rejects.toThrow(RequestError);
 
                 if (cluster.checkIfServerVersionLessThan("7.0.0")) {
                     await expect(
-                        client.bitcount(
-                            key1,
-                            new BitOffsetOptions(2, 5, BitmapIndexType.BIT),
-                        ),
+                        client.bitcount(key1, {
+                            start: 2,
+                            end: 5,
+                            indexType: BitmapIndexType.BIT,
+                        }),
                     ).rejects.toThrow();
                     await expect(
-                        client.bitcount(
-                            key1,
-                            new BitOffsetOptions(2, 5, BitmapIndexType.BYTE),
-                        ),
+                        client.bitcount(key1, {
+                            start: 2,
+                            end: 5,
+                            indexType: BitmapIndexType.BYTE,
+                        }),
                     ).rejects.toThrow();
                 } else {
                     expect(
-                        await client.bitcount(
-                            key1,
-                            new BitOffsetOptions(2, 5, BitmapIndexType.BYTE),
-                        ),
+                        await client.bitcount(key1, {
+                            start: 2,
+                            end: 5,
+                            indexType: BitmapIndexType.BYTE,
+                        }),
                     ).toEqual(16);
                     expect(
-                        await client.bitcount(
-                            key1,
-                            new BitOffsetOptions(5, 30, BitmapIndexType.BIT),
-                        ),
+                        await client.bitcount(key1, {
+                            start: 5,
+                            end: 30,
+                            indexType: BitmapIndexType.BIT,
+                        }),
                     ).toEqual(17);
                     expect(
-                        await client.bitcount(
-                            key1,
-                            new BitOffsetOptions(5, -5, BitmapIndexType.BIT),
-                        ),
+                        await client.bitcount(key1, {
+                            start: 5,
+                            end: -5,
+                            indexType: BitmapIndexType.BIT,
+                        }),
                     ).toEqual(23);
                     expect(
-                        await client.bitcount(
-                            uuidv4(),
-                            new BitOffsetOptions(2, 5, BitmapIndexType.BYTE),
-                        ),
+                        await client.bitcount(uuidv4(), {
+                            start: 2,
+                            end: 5,
+                            indexType: BitmapIndexType.BYTE,
+                        }),
                     ).toEqual(0);
                     // key exists, but it is not a string
                     await expect(
-                        client.bitcount(
-                            key2,
-                            new BitOffsetOptions(1, 1, BitmapIndexType.BYTE),
-                        ),
+                        client.bitcount(key2, {
+                            start: 1,
+                            end: 1,
+                            indexType: BitmapIndexType.BYTE,
+                        }),
                     ).rejects.toThrow(RequestError);
                 }
             }, protocol);
@@ -4591,14 +4799,14 @@ export function runBaseTests<Context>(config: {
                 const key1 = uuidv4();
                 const key2 = uuidv4();
                 const membersToCoordinates = new Map<string, GeospatialData>();
-                membersToCoordinates.set(
-                    "Palermo",
-                    new GeospatialData(13.361389, 38.115556),
-                );
-                membersToCoordinates.set(
-                    "Catania",
-                    new GeospatialData(15.087269, 37.502669),
-                );
+                membersToCoordinates.set("Palermo", {
+                    longitude: 13.361389,
+                    latitude: 38.115556,
+                });
+                membersToCoordinates.set("Catania", {
+                    longitude: 15.087269,
+                    latitude: 37.502669,
+                });
 
                 // default geoadd
                 expect(await client.geoadd(key1, membersToCoordinates)).toBe(2);
@@ -4626,45 +4834,34 @@ export function runBaseTests<Context>(config: {
                 expect(geopos).toEqual([null]);
 
                 // with update mode options
-                membersToCoordinates.set(
-                    "Catania",
-                    new GeospatialData(15.087269, 39),
-                );
+                membersToCoordinates.set("Catania", {
+                    longitude: 15.087269,
+                    latitude: 39,
+                });
                 expect(
-                    await client.geoadd(
-                        key1,
-                        membersToCoordinates,
-                        new GeoAddOptions({
-                            updateMode:
-                                ConditionalChange.ONLY_IF_DOES_NOT_EXIST,
-                        }),
-                    ),
+                    await client.geoadd(key1, membersToCoordinates, {
+                        updateMode: ConditionalChange.ONLY_IF_DOES_NOT_EXIST,
+                    }),
                 ).toBe(0);
                 expect(
-                    await client.geoadd(
-                        key1,
-                        membersToCoordinates,
-                        new GeoAddOptions({
-                            updateMode: ConditionalChange.ONLY_IF_EXISTS,
-                        }),
-                    ),
+                    await client.geoadd(key1, membersToCoordinates, {
+                        updateMode: ConditionalChange.ONLY_IF_EXISTS,
+                    }),
                 ).toBe(0);
 
                 // with changed option
-                membersToCoordinates.set(
-                    "Catania",
-                    new GeospatialData(15.087269, 40),
-                );
-                membersToCoordinates.set(
-                    "Tel-Aviv",
-                    new GeospatialData(32.0853, 34.7818),
-                );
+                membersToCoordinates.set("Catania", {
+                    longitude: 15.087269,
+                    latitude: 40,
+                });
+                membersToCoordinates.set("Tel-Aviv", {
+                    longitude: 32.0853,
+                    latitude: 34.7818,
+                });
                 expect(
-                    await client.geoadd(
-                        key1,
-                        membersToCoordinates,
-                        new GeoAddOptions({ changed: true }),
-                    ),
+                    await client.geoadd(key1, membersToCoordinates, {
+                        changed: true,
+                    }),
                 ).toBe(2);
 
                 // key exists but holding non-zset value
@@ -4691,25 +4888,25 @@ export function runBaseTests<Context>(config: {
                 await expect(
                     client.geoadd(
                         key,
-                        new Map([["Place", new GeospatialData(-181, 0)]]),
+                        new Map([["Place", { longitude: -181, latitude: 0 }]]),
                     ),
                 ).rejects.toThrow();
                 await expect(
                     client.geoadd(
                         key,
-                        new Map([["Place", new GeospatialData(181, 0)]]),
+                        new Map([["Place", { longitude: 181, latitude: 0 }]]),
                     ),
                 ).rejects.toThrow();
                 await expect(
                     client.geoadd(
                         key,
-                        new Map([["Place", new GeospatialData(0, 86)]]),
+                        new Map([["Place", { longitude: 0, latitude: 86 }]]),
                     ),
                 ).rejects.toThrow();
                 await expect(
                     client.geoadd(
                         key,
-                        new Map([["Place", new GeospatialData(0, -86)]]),
+                        new Map([["Place", { longitude: 0, latitude: -86 }]]),
                     ),
                 ).rejects.toThrow();
             }, protocol);
@@ -4783,6 +4980,125 @@ export function runBaseTests<Context>(config: {
     );
 
     it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `zincrby test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const key = "{key}" + uuidv4();
+                const member = "{member}-1" + uuidv4();
+                const othermember = "{member}-1" + uuidv4();
+                const stringKey = "{key}-string" + uuidv4();
+
+                // key does not exist
+                expect(await client.zincrby(key, 2.5, member)).toEqual(2.5);
+                expect(await client.zscore(key, member)).toEqual(2.5);
+
+                // key exists, but value doesn't
+                expect(await client.zincrby(key, -3.3, othermember)).toEqual(
+                    -3.3,
+                );
+                expect(await client.zscore(key, othermember)).toEqual(-3.3);
+
+                // updating existing value in existing key
+                expect(await client.zincrby(key, 1.0, member)).toEqual(3.5);
+                expect(await client.zscore(key, member)).toEqual(3.5);
+
+                // Key exists, but it is not a sorted set
+                expect(await client.set(stringKey, "value")).toEqual("OK");
+                await expect(
+                    client.zincrby(stringKey, 0.5, "_"),
+                ).rejects.toThrow(RequestError);
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `bzmpop test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient, cluster: RedisCluster) => {
+                if (cluster.checkIfServerVersionLessThan("7.0.0")) return;
+                const key1 = "{key}-1" + uuidv4();
+                const key2 = "{key}-2" + uuidv4();
+                const nonExistingKey = "{key}-0" + uuidv4();
+                const stringKey = "{key}-string" + uuidv4();
+
+                expect(await client.zadd(key1, { a1: 1, b1: 2 })).toEqual(2);
+                expect(await client.zadd(key2, { a2: 0.1, b2: 0.2 })).toEqual(
+                    2,
+                );
+
+                checkSimple(
+                    await client.bzmpop([key1, key2], ScoreFilter.MAX, 0.1),
+                ).toEqual([key1, { b1: 2 }]);
+                checkSimple(
+                    await client.bzmpop([key2, key1], ScoreFilter.MAX, 0.1, 10),
+                ).toEqual([key2, { a2: 0.1, b2: 0.2 }]);
+
+                // ensure that command doesn't time out even if timeout > request timeout (250ms by default)
+                expect(
+                    await client.bzmpop([nonExistingKey], ScoreFilter.MAX, 0.5),
+                ).toBeNull;
+                expect(
+                    await client.bzmpop(
+                        [nonExistingKey],
+                        ScoreFilter.MAX,
+                        0.55,
+                        1,
+                    ),
+                ).toBeNull;
+
+                // key exists, but it is not a sorted set
+                expect(await client.set(stringKey, "value")).toEqual("OK");
+                await expect(
+                    client.bzmpop([stringKey], ScoreFilter.MAX, 0.1),
+                ).rejects.toThrow(RequestError);
+                await expect(
+                    client.bzmpop([stringKey], ScoreFilter.MAX, 0.1, 1),
+                ).rejects.toThrow(RequestError);
+
+                // incorrect argument: key list should not be empty
+                await expect(
+                    client.bzmpop([], ScoreFilter.MAX, 0.1, 1),
+                ).rejects.toThrow(RequestError);
+
+                // incorrect argument: count should be greater than 0
+                await expect(
+                    client.bzmpop([key1], ScoreFilter.MAX, 0.1, 0),
+                ).rejects.toThrow(RequestError);
+
+                // incorrect argument: timeout can not be a negative number
+                await expect(
+                    client.bzmpop([key1], ScoreFilter.MAX, -1, 10),
+                ).rejects.toThrow(RequestError);
+
+                // check that order of entries in the response is preserved
+                const entries: Record<string, number> = {};
+
+                for (let i = 0; i < 10; i++) {
+                    // a0 => 0, a1 => 1 etc
+                    entries["a" + i] = i;
+                }
+
+                expect(await client.zadd(key2, entries)).toEqual(10);
+                const result = await client.bzmpop(
+                    [key2],
+                    ScoreFilter.MIN,
+                    0.1,
+                    10,
+                );
+
+                if (result) {
+                    expect(result[1]).toEqual(entries);
+                }
+
+                // TODO: add test case with 0 timeout (no timeout) should never time out,
+                // but we wrap the test with timeout to avoid test failing or stuck forever
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
         `geodist test_%p`,
         async (protocol) => {
             await runTest(async (client: BaseClient) => {
@@ -4797,14 +5113,14 @@ export function runBaseTests<Context>(config: {
 
                 // adding the geo locations
                 const membersToCoordinates = new Map<string, GeospatialData>();
-                membersToCoordinates.set(
-                    member1,
-                    new GeospatialData(13.361389, 38.115556),
-                );
-                membersToCoordinates.set(
-                    member2,
-                    new GeospatialData(15.087269, 37.502669),
-                );
+                membersToCoordinates.set(member1, {
+                    longitude: 13.361389,
+                    latitude: 38.115556,
+                });
+                membersToCoordinates.set(member2, {
+                    longitude: 15.087269,
+                    latitude: 37.502669,
+                });
                 expect(await client.geoadd(key1, membersToCoordinates)).toBe(2);
 
                 // checking result with default metric
@@ -4849,14 +5165,14 @@ export function runBaseTests<Context>(config: {
 
                 // adding the geo locations
                 const membersToCoordinates = new Map<string, GeospatialData>();
-                membersToCoordinates.set(
-                    "Palermo",
-                    new GeospatialData(13.361389, 38.115556),
-                );
-                membersToCoordinates.set(
-                    "Catania",
-                    new GeospatialData(15.087269, 37.502669),
-                );
+                membersToCoordinates.set("Palermo", {
+                    longitude: 13.361389,
+                    latitude: 38.115556,
+                });
+                membersToCoordinates.set("Catania", {
+                    longitude: 15.087269,
+                    latitude: 37.502669,
+                });
                 expect(await client.geoadd(key1, membersToCoordinates)).toBe(2);
 
                 // checking result with default metric